move detectors to .mts extension
