+++ packages/detectors/esbuild.ts
@@ -1,12 +1,17 @@
 import { build } from 'esbuild'
-import { resolve } from 'path'
+import { dirname, resolve } from 'path'
 import glob from 'tiny-glob/sync.js'
+import { fileURLToPath } from 'url'
+
+//@ts-ignore
+const __filename = fileURLToPath(import.meta.url)
+const __dirname = dirname(__filename)
 
 const watch = process.argv.includes('--watch')
 
 const getPath = (file: string) => resolve(__dirname, file)
 
-const files = glob('./src/**/*.*ts')
+const files = glob('./src/**/*.*mts')
 	.map((file) =>
 		resolve(file)
 			.substring(resolve(__dirname, `./src/`).length + 1)
@@ -21,7 +26,7 @@ files.forEach((file) =>
 		build({
 			entryPoints: [getPath(`./src/${file}`)],
 			bundle: true,
-			outfile: getPath(`../../detectors/${file.replace('.ts', `.${format === 'esm' ? 'm' : 'c'}js`)}`),
+			outfile: getPath(`../../detectors/${file.replace('.mts', `.${format === 'esm' ? 'm' : 'c'}js`)}`),
 			platform: 'browser',
 			format,
 			minify: true,
