+++ packages/detectors/package.json
@@ -15,5 +15,6 @@
     "typescript": "^4.7.4",
     "uvu": "^0.5.6",
     "watchlist": "^0.3.1"
-  }
+  },
+  "type": "module"
 }
