+++ packages/detectors/src/detect.mts
