+++ packages/detectors/src/detectors/browser/navigator.mts
@@ -1,4 +1,4 @@
 import type { Locale } from '@typesafe-i18n/runtime/core.mjs'
-import type { LocaleDetector } from '../../detect'
+import type { LocaleDetector } from '../../detect.mjs'
 
 export const navigatorDetector: LocaleDetector = (): Locale[] => (navigator?.languages as Locale[]) || []
