+++ packages/detectors/src/detectors/browser/query-string.mts
@@ -1,5 +1,5 @@
 import type { Locale } from '@typesafe-i18n/runtime/core.mjs'
-import type { LocaleDetector } from '../../detect'
+import type { LocaleDetector } from '../../detect.mjs'
 
 export const initQueryStringDetector =
 	(key = 'lang'): LocaleDetector =>
