+++ packages/detectors/src/detectors/server/accept-language-header.mts
@@ -1,6 +1,6 @@
 import type { Locale } from '@typesafe-i18n/runtime/core.mjs'
 import { isNotEmpty } from 'typesafe-utils'
-import type { LocaleDetector } from '../../detect'
+import type { LocaleDetector } from '../../detect.mjs'
 
 type ObjectWithHeaders = { headers: Record<string, string | string[] | undefined> }
