+++ packages/detectors/src/detectors/server/request-hostname.mts
@@ -1,5 +1,5 @@
 import type { Locale } from '@typesafe-i18n/runtime/core.mjs'
-import type { LocaleDetector } from '../../detect'
+import type { LocaleDetector } from '../../detect.mjs'
 
 type Request = {
 	hostname: string
