--- packages/detectors/src/index.ts
@@ -1,12 +1,0 @@
-export { detectLocale } from './detect'
-export type { LocaleDetector } from './detect'
-export { documentCookieDetector, initDocumentCookieDetector } from './detectors/browser/document-cookie'
-export { htmlLangAttributeDetector } from './detectors/browser/html-lang-attribute'
-export { initLocalStorageDetector, localStorageDetector } from './detectors/browser/localstorage-detector'
-export { navigatorDetector } from './detectors/browser/navigator'
-export { initQueryStringDetector, queryStringDetector } from './detectors/browser/query-string'
-export { initSessionStorageDetector, sessionStorageDetector } from './detectors/browser/sessionstorage-detector'
-export { initAcceptLanguageHeaderDetector } from './detectors/server/accept-language-header'
-export { initRequestCookiesDetector } from './detectors/server/request-cookies'
-export { initRequestHostnameDetector } from './detectors/server/request-hostname'
-export { initRequestParametersDetector } from './detectors/server/request-parameters'
