+++ packages/detectors/test/server/request-hostname.test.ts
@@ -1,7 +1,7 @@
 import type { Locale } from '@typesafe-i18n/runtime/core.mjs'
 import { suite } from 'uvu'
 import * as assert from 'uvu/assert'
-import { initRequestHostnameDetector } from '../../src/detectors/server/request-hostname'
+import { initRequestHostnameDetector } from '../../src/detectors/server/request-hostname.mjs'
 
 const test = suite('detector:request-hostname')
