+++ packages/detectors/test/server/request-parameters.test.ts
@@ -1,7 +1,7 @@
 import type { Locale } from '@typesafe-i18n/runtime/core.mjs'
 import { suite } from 'uvu'
 import * as assert from 'uvu/assert'
-import { initRequestParametersDetector } from '../../src/detectors/server/request-parameters'
+import { initRequestParametersDetector } from '../../src/detectors/server/request-parameters.mjs'
 
 const test = suite('detector:request-parameters')
